# Conflicts:
#	src/Orc.Skia.Example.NET/Orc.Skia.Example.NET.csproj
#	src/Orc.Skia/FodyWeavers.xml
#	src/Orc.Skia/Orc.Skia.csproj
#	src/Orc.Skia/costura32/libSkiaSharp.dll
#	src/Orc.Skia/costura64/libSkiaSharp.dll

+++ deployment/cake/components-tasks.cake
@@ -1,7 +1,5 @@
 #l "components-variables.cake"
 
-#addin "nuget:?package=Cake.FileHelpers&version=3.0.0"
-
 using System.Xml.Linq;
 
 //-------------------------------------------------------------
@@ -52,7 +50,8 @@ public class ComponentsProcessor : ProcessorBase
         {
             foreach (var component in BuildContext.Components.Items)
             {
-                var cacheDirectory = Environment.ExpandEnvironmentVariables(string.Format("%userprofile%/.nuget/packages/{0}/{1}", component, BuildContext.General.Version.NuGet));
+                var expandableCacheDirectory = System.IO.Path.Combine("%userprofile%", ".nuget", "packages", component, BuildContext.General.Version.NuGet);
+                var cacheDirectory = Environment.ExpandEnvironmentVariables(expandableCacheDirectory);
 
                 CakeContext.Information("Checking for existing local NuGet cached version at '{0}'", cacheDirectory);
 
@@ -124,13 +123,14 @@ public class ComponentsProcessor : ProcessorBase
                 PlatformTarget = PlatformTarget.MSIL
             };
 
-            ConfigureMsBuild(BuildContext, msBuildSettings, component);
+            ConfigureMsBuild(BuildContext, msBuildSettings, component, "build");
             
             // Note: we need to set OverridableOutputPath because we need to be able to respect
             // AppendTargetFrameworkToOutputPath which isn't possible for global properties (which
             // are properties passed in using the command line)
             var outputDirectory = GetProjectOutputDirectory(BuildContext, component);
             CakeContext.Information("Output directory: '{0}'", outputDirectory);
+            msBuildSettings.WithProperty("OverridableOutputRootPath", BuildContext.General.OutputRootDirectory);
             msBuildSettings.WithProperty("OverridableOutputPath", outputDirectory);
             msBuildSettings.WithProperty("PackageOutputPath", BuildContext.General.OutputRootDirectory);
 
@@ -139,27 +139,23 @@ public class ComponentsProcessor : ProcessorBase
             {
                 var repositoryUrl = BuildContext.General.Repository.Url;
                 var repositoryCommitId = BuildContext.General.Repository.CommitId;
-                if (!BuildContext.General.SourceLink.IsDisabled && 
-                    !BuildContext.General.IsLocalBuild && 
-                    !string.IsNullOrWhiteSpace(repositoryUrl))
-                {       
-                    CakeContext.Information("Repository url is specified, enabling SourceLink to commit '{0}/commit/{1}'", 
-                        repositoryUrl, repositoryCommitId);
 
-                    // TODO: For now we are assuming everything is git, we might need to change that in the future
-                    // See why we set the values at https://github.com/dotnet/sourcelink/issues/159#issuecomment-427639278
-                    msBuildSettings.WithProperty("EnableSourceLink", "true");
-                    msBuildSettings.WithProperty("EnableSourceControlManagerQueries", "false");
-                    msBuildSettings.WithProperty("PublishRepositoryUrl", "true");
-                    msBuildSettings.WithProperty("RepositoryType", "git");
-                    msBuildSettings.WithProperty("RepositoryUrl", repositoryUrl);
-                    msBuildSettings.WithProperty("RevisionId", repositoryCommitId);
+                CakeContext.Information("Repository url is specified, enabling SourceLink to commit '{0}/commit/{1}'", 
+                    repositoryUrl, repositoryCommitId);
 
-                    InjectSourceLinkInProjectFile(BuildContext, projectFileName);
-                }
+                // TODO: For now we are assuming everything is git, we might need to change that in the future
+                // See why we set the values at https://github.com/dotnet/sourcelink/issues/159#issuecomment-427639278
+                msBuildSettings.WithProperty("EnableSourceLink", "true");
+                msBuildSettings.WithProperty("EnableSourceControlManagerQueries", "false");
+                msBuildSettings.WithProperty("PublishRepositoryUrl", "true");
+                msBuildSettings.WithProperty("RepositoryType", "git");
+                msBuildSettings.WithProperty("RepositoryUrl", repositoryUrl);
+                msBuildSettings.WithProperty("RevisionId", repositoryCommitId);
+
+                InjectSourceLinkInProjectFile(BuildContext, projectFileName);
             }
 
-            CakeContext.MSBuild(projectFileName, msBuildSettings);
+            RunMsBuild(BuildContext, component, projectFileName, msBuildSettings, "build");
         }        
     }
 
@@ -174,10 +170,23 @@ public class ComponentsProcessor : ProcessorBase
 
         foreach (var component in BuildContext.Components.Items)
         {
+            // Note: some projects, such as Catel.Fody, require packaging
+            // of non-deployable projects
+            if (BuildContext.General.SkipComponentsThatAreNotDeployable && 
+                !ShouldDeployProject(BuildContext, component))
+            {
+                CakeContext.Information("Component '{0}' should not be deployed", component);
+                continue;
+            }
+
+
+            // Special exception for Blazor projects
+            var isBlazorProject = IsBlazorProject(BuildContext, component);
+
             BuildContext.CakeContext.LogSeparator("Packaging component '{0}'", component);
 
-            var projectDirectory = string.Format("./src/{0}", component);
-            var projectFileName = string.Format("{0}/{1}.csproj", projectDirectory, component);
+            var projectDirectory = GetProjectDirectory(component);
+            var projectFileName = GetProjectFileName(BuildContext, component);
             var outputDirectory = GetProjectOutputDirectory(BuildContext, component);
             CakeContext.Information("Output directory: '{0}'", outputDirectory);
 
@@ -194,19 +203,23 @@ public class ComponentsProcessor : ProcessorBase
             var binFiles = CakeContext.GetFiles(binFolderPattern);
             CakeContext.DeleteFiles(binFiles);
 
-            var objFolderPattern = string.Format("{0}/obj/{1}/**.dll", projectDirectory, configurationName);
+            if (!isBlazorProject)
+            {
+                var objFolderPattern = string.Format("{0}/obj/{1}/**.dll", projectDirectory, configurationName);
 
-            CakeContext.Information("Deleting 'bin' directory contents using '{0}'", objFolderPattern);
+                CakeContext.Information("Deleting 'bin' directory contents using '{0}'", objFolderPattern);
 
-            var objFiles = CakeContext.GetFiles(objFolderPattern);
-            CakeContext.DeleteFiles(objFiles);
+                var objFiles = CakeContext.GetFiles(objFolderPattern);
+                CakeContext.DeleteFiles(objFiles);
+            }
 
             CakeContext.Information(string.Empty);
 
             // Step 2: Go packaging!
             CakeContext.Information("Using 'msbuild' to package '{0}'", component);
 
-            var msBuildSettings = new MSBuildSettings {
+            var msBuildSettings = new MSBuildSettings 
+            {
                 Verbosity = Verbosity.Quiet,
                 //Verbosity = Verbosity.Diagnostic,
                 ToolVersion = MSBuildToolVersion.Default,
@@ -220,6 +233,7 @@ public class ComponentsProcessor : ProcessorBase
             // Note: we need to set OverridableOutputPath because we need to be able to respect
             // AppendTargetFrameworkToOutputPath which isn't possible for global properties (which
             // are properties passed in using the command line)
+            msBuildSettings.WithProperty("OverridableOutputRootPath", BuildContext.General.OutputRootDirectory);
             msBuildSettings.WithProperty("OverridableOutputPath", outputDirectory);
             msBuildSettings.WithProperty("PackageOutputPath", BuildContext.General.OutputRootDirectory);
             msBuildSettings.WithProperty("ConfigurationName", configurationName);
@@ -242,14 +256,28 @@ public class ComponentsProcessor : ProcessorBase
                 msBuildSettings.WithProperty("RevisionId", repositoryCommitId);
             }
             
+            // Disable Multilingual App Toolkit (MAT) during packaging
+            msBuildSettings.WithProperty("DisableMAT", "true");
+
             // Fix for .NET Core 3.0, see https://github.com/dotnet/core-sdk/issues/192, it
             // uses obj/release instead of [outputdirectory]
             msBuildSettings.WithProperty("DotNetPackIntermediateOutputPath", outputDirectory);
-            
-            msBuildSettings.WithProperty("NoBuild", "true");
+
+            var noBuild = true;
+
+            if (isBlazorProject)
+            {
+                CakeContext.Information("Allowing build and package restore during package phase since this is a Blazor project which requires the 'obj' directory");
+
+                msBuildSettings.WithProperty("ResolveNuGetPackages", "true");
+                msBuildSettings.Restore = true;
+                noBuild = false;
+            }
+
+            msBuildSettings.WithProperty("NoBuild", noBuild.ToString());
             msBuildSettings.Targets.Add("Pack");
 
-            CakeContext.MSBuild(projectFileName, msBuildSettings);
+            RunMsBuild(BuildContext, component, projectFileName, msBuildSettings, "pack");
 
             BuildContext.CakeContext.LogSeparator();
         }
@@ -261,16 +289,15 @@ public class ComponentsProcessor : ProcessorBase
         {
             // For details, see https://docs.microsoft.com/en-us/nuget/create-packages/sign-a-package
             // nuget sign MyPackage.nupkg -CertificateSubjectName <MyCertSubjectName> -Timestamper <TimestampServiceURL>
-            var filesToSign = CakeContext.GetFiles(string.Format("{0}/*.nupkg", BuildContext.General.OutputRootDirectory));
-
+            var filesToSign = CakeContext.GetFiles($"{BuildContext.General.OutputRootDirectory}/*.nupkg");
+            
             foreach (var fileToSign in filesToSign)
             {
-                CakeContext.Information("Signing NuGet package '{0}' using certificate subject '{1}'", fileToSign, BuildContext.General.CodeSign.CertificateSubjectName);
+                CakeContext.Information($"Signing NuGet package '{fileToSign}' using certificate subject '{BuildContext.General.CodeSign.CertificateSubjectName}'");
 
                 var exitCode = CakeContext.StartProcess(BuildContext.General.NuGet.Executable, new ProcessSettings
                 {
-                    Arguments = string.Format("sign \"{0}\" -CertificateSubjectName \"{1}\" -Timestamper \"{2}\"", 
-                        fileToSign, BuildContext.General.CodeSign.CertificateSubjectName, BuildContext.General.CodeSign.TimeStampUri)
+                    Arguments = $"sign \"{fileToSign}\" -CertificateSubjectName \"{BuildContext.General.CodeSign.CertificateSubjectName}\" -Timestamper \"{BuildContext.General.CodeSign.TimeStampUri}\""
                 });
 
                 CakeContext.Information("Signing NuGet package exited with '{0}'", exitCode);
@@ -295,7 +322,7 @@ public class ComponentsProcessor : ProcessorBase
 
             BuildContext.CakeContext.LogSeparator("Deploying component '{0}'", component);
 
-            var packageToPush = string.Format("{0}/{1}.{2}.nupkg", BuildContext.General.OutputRootDirectory, component, BuildContext.General.Version.NuGet);
+            var packageToPush = System.IO.Path.Combine(BuildContext.General.OutputRootDirectory, $"{component}.{BuildContext.General.Version.NuGet}.nupkg");
             var nuGetRepositoryUrl = GetComponentNuGetRepositoryUrl(component);
             var nuGetRepositoryApiKey = GetComponentNuGetRepositoryApiKey(component);
